AppVeyor: install and include OpenSSL
Qt5 does not include OpenSSL libraries it was build against. See https://forum.qt.io/topic/95700/qsslsocket-tls-initialization-failed/18

As a result we have to make sure a compatible libssl.dll and libcrypto.dll is included in PATH when the software is loaded. It seems not to be detected by our Python dependency script. So, I added it manually like the JACK library.

Important: when upgrading/downgrading the Qt version a different version of OpenSSL might be required. One has to manually check the version it was build against https://wiki.qt.io/Qt_5.15_Tools_and_Versions#windows_10_x86_64_7.

But this can also be abstracted by a unit test as soon as the GUI can also be covered (I think we have to build a dedicated library as in https://github.com/hydrogen-music/hydrogen/pull/1329 first).