AppVeyor: use Qt 5.15.2-14
for some reason mingw-w64-x86_64-qt5-base-5.15.3+kde+r174-2-any was installed as part of mingw-w64-x86_64-qt5-5.15.2-15-any. Not quite sure how this version could be called 5.15.2-15 but it is still the one with the bug we want to circumvent. On my Windows machine I think this behaved differently.

Let us try 5.12.2-14 instead.